requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
base58==2.1.1
bitarray==2.9.2
bokeh==3.4.1
cachetools==5.5.0
ccxt==4.4.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==42.0.5
cycler==0.12.1
cytoolz==1.0.0
db-dtypes==1.3.0
debugpy==1.8.6
decorator==5.1.1
docstring_parser==0.16
emp-orderly==1.0
emp-orderly-types==1.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-rpc-py==0.1.8.post6
eth-typeshed-py==0.1.9.post4
eth-typing==4.1.0
eth-utils==4.1.0
eth_abi==5.1.0
executing==2.1.0
fonttools==4.54.1
frozenlist==1.4.1
google-api-core==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pandas-gbq==0.23.2
pandas_ta==0.3.14b0
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.4.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydata-google-auth==1.8.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.31.0
requests-oauthlib==2.0.0
rlp==4.0.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
threadpoolctl==3.5.0
toolz==1.0.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
xyzservices==2024.9.0
yarl==1.13.1
